Moved Scene and LevelEditorScene to scenes package
