+++ src/main/java/scenes/LevelEditorScene.java
@@ -1,4 +1,4 @@
-package jade;
+package scenes;
 
 import components.Rigidbody;
 import components.Sprite;
@@ -6,11 +6,16 @@ import components.SpriteRenderer;
 import components.Spritesheet;
 import imgui.ImGui;
 import imgui.ImVec2;
+import jade.Camera;
+import jade.GameObject;
+import jade.MouseListener;
+import jade.Transform;
 import org.joml.Vector2f;
 import org.joml.Vector4f;
+import scenes.Scene;
 import util.AssetPool;
 
-public class LevelEditorScene extends Scene{
+public class LevelEditorScene extends Scene {
 
     private GameObject obj1;
     private Spritesheet sprites;
