+++ src/main/java/scenes/LevelScene.java
@@ -1,6 +1,9 @@
-package jade;
+package scenes;
 
-public class LevelScene extends Scene{
+import jade.Window;
+import scenes.Scene;
+
+public class LevelScene extends Scene {
     public LevelScene() {
         System.out.println("Inside level scene");
         Window.get().r = 1;
